requirements.txt
absl-py==0.2.2
alabaster==0.7.10
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.3
astropy==3.0.1
atomicwrites==1.2.1
attrs==17.4.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
blinker==1.4
boto==2.48.0
Bottleneck==1.2.1
brewer2mpl==1.4.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
colorama==0.3.9
comtypes==1.1.4
contextlib2==0.5.5
cryptography==2.2.1
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.2
datashape==0.5.4
decorator==4.3.0
distributed==1.21.6
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-Dance==1.1.0
Flask-SeaSurf==0.2.2
gast==0.2.0
gevent==1.3a1
ggplot==0.11.2
glob2==0.5
greenlet==0.4.13
grpcio==1.12.1
gunicorn==19.9.0
h5py==2.7.1
heapdict==1.0.0
html2text==2018.1.9
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
imageio==2.3.0
imagesize==1.0.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
kiwisolver==1.0.1
lazy==1.3
lazy-object-proxy==1.3.1
llvmlite==0.22.0
locket==0.2.0
lxml==4.2.1
mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
more-itertools==4.1.0
msgpack-python==0.5.6
multipledispatch==0.5.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.1
numba==0.37.0
numexpr==2.6.4
numpy==1.15.2
numpydoc==0.8.0
oauth2client==4.1.2
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.23.4
pandas-datareader==0.6.0
pandocfilters==1.4.1
parso==0.2.0
partd==0.3.8
path.py==8.2.1
pathlib2==2.3.2
patsy==0.5.0
pbr==4.0.4
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.2
plotly==2.5.1
pluggy==0.7.1
ply==3.11
progressbar==2.5
prompt-toolkit==1.0.15
protobuf==3.5.2
psutil==5.4.5
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pygpu==0.7.6
PyJWT==1.6.4
pylint==1.8.4
PyMySQL==0.9.2
pyodbc==4.0.17
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.2.1
pytest-doctestplus==0.1.2
pytest-openfiles==0.2.0
pytest-remotedata==0.2.0
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pywin32==221
pywinpty==0.5
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.0.0
retrying==1.3.3
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.9
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.11
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
traitlets==4.3.2
tweepy==3.5.0
typing==3.6.4
ua-parser==0.8.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
URLObject==2.4.3
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wptools==0.4.15
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.7
xlwt==1.3.0
zict==0.1.3
